[IMP] firmware: Application 1.3.3 * Resolve integer overflow bug for long metadata * Resolve issue with title lengths in BMBT menus * Update to toolchain 2.10

+++ firmware/application/lib/bt/bt_bm83.c
@@ -1113,7 +1113,7 @@ void BM83ProcessDataGetAllAttributes(
     BT_t *bt,
     uint8_t *data,
     uint8_t attributeCount,
-    uint8_t bytePos
+    uint16_t bytePos
 ) {
     uint8_t dataDiffers = 0;
     bt->metadataStatus = BT_METADATA_STATUS_CUR;
